Split file into multiple separate files
