+++ patches/tModLoader/Terraria/ModLoader/ConversionRunCodeValues.cs
@@ -1,0 +1,8 @@
+﻿namespace Terraria.ModLoader;
+
+public enum ConversionRunCodeValues : sbyte
+{
+	DontRun = -2,
+	Run = 0,
+	Break = 1,
+}
